Feature: Scene Debug
"Scene Debug" is a WIP tool for getting certain info and stats about the current Blend file.

As first experiment, I added an operator to list certain (cycles) shader nodes being used in the scene. This came in handy while looking on why a Cycles shot was super noisy, found out via this tool that there were many Emission objects that contributed almost no light to the scene, but lots of noise. Same for looking if there are SSS materials, or Glossy, Glass, and so on.

Find it on the Scene Debug panel under Scene properties.

I plan on adding more tools like this to help debug issues and find out stuff about .blend files. Perhaps not super useful on simple setups but on big production files it comes in handy.