Feature: Object ID for objects inside DupliGroups
Set an ID for objects inside dupligroups. Just loop through all the objects in a dupligroup and set the Empty (or object acting as dupli)'s object ID instead.

In order for the OB ID to be loaded afterwards on computers without Amaranth installed, it will create a text file (called AmaranthStartup.py) inside the .blend that will run on startup and set the OB IDs, if any. The way this is implemented might be a bit hacky but I couldn't find on another reliable way.

Fix: Library display for duplis was printing an error when there was no library.