Feature: Set Render Samples
When lighting on Cycles, usually you have a sweet number of samples for a noise-free render. If you want to quickly render a preview you have to change this value and remember the previous value.
The "Set Render Samples" button in the Sampling panel, will save this number and let you choose different percentages to quickly go up/down in samples number, while being able to go back to the final number.