New: Set Scene Layers to be enabled for final render
When doing rendering, you have a set of layers that should be enabled for a final render to look properly.
Sometimes you change the layers by mistake or for testing or just a lot of time passed since you
opened the file last.

This feature lets you save in the .blend file the Scene layers that should be enabled for final renders,
so you can set it when you're sure and forget about it. Change the layers anytime and swap back to
how it should be setup for a proper render.

You can even attach this operator to your farm script or so for safer renders :)

How it looks at the moment: http://pasteall.org/pic/show.php?id=69930

Hit the Save Current Layers.. button first, then it becomes "Replace" if there's already a setup.
Below there's a button to view the layers for rendering.
You can also turn on/off individual layers from the setup by clicking on them.

It will help me a lot on my day by day lighting work. Hope it helps you too!