Feature: Jump to in-between keyframes
With Up/Down key you jump to next/prev keyframes. This commit adds the posibility to jump to the frame in-between keyframes of the active object. Use: Ctrl+Shift Up/Down.

So if you have a keyframe in frame 1, and 100. Ctrl+Shift+Up will take you to frame 50.

is_keyframe function based on a reply by CoDEmanX on blender.stackexchange, thanks!