New Debug feature in progress: List Users for Datablock
This is a work in progress feature that lists all the users for a specific datablock.

I needed it to debug some shots in gooseberry. There were images in the .blend that
I had no idea where they were coming from. So this is a way to help solve that.

This is work in progress as I plan on adding more info and types to search from
(only works with images at the moment because that's what I needed),
I just pushed this so I can keep working at home.